Arreglo de visualizacion de back end
