+++ Proyect_DSW/controllers/cifrado_AES.py
@@ -1,0 +1,33 @@
+from cryptography.hazmat.primitives.kdf.hkdf import HKDF
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
+import Proyect_DSW.controllers.keys as keys
+
+#========= Funciones para cifrar AES =====================
+def generar_llave_aes(password):
+    password = password.encode('utf-8')
+    derived_key = HKDF(algorithm=hashes.SHA256(),
+                       length=32,
+                       salt=None,
+                       info=b'handshake data ',
+                       backend=default_backend()).derive(password)
+    return derived_key
+
+
+def cifrar(key, llave_aes, iv):
+    key = keys.convertir_llave_privada_bytes(key)
+    aesCipher = Cipher(algorithms.AES(llave_aes), modes.CTR(iv),
+                       backend=default_backend())
+    cifrador = aesCipher.encryptor()
+    cifrado = cifrador.update(key)
+    cifrador.finalize()
+    return cifrado
+
+def descifrar(cifrado, llave_aes, iv):
+    aesCipher = Cipher(algorithms.AES(llave_aes), modes.CTR(iv),
+                       backend=default_backend())
+    descifrador = aesCipher.decryptor()
+    plano = descifrador.update(cifrado)
+    descifrador.finalize()
+    return plano
