+++ Proyect_DSW/controllers/keys.py
@@ -1,0 +1,51 @@
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric import ec
+import sys
+
+#========= Funciones para generar llaves =====================
+def generar_llave_privada():
+    private_key = ec.generate_private_key(ec.SECP384R1(), default_backend())
+    return private_key
+
+
+def generar_llave_publica(llave_privada):
+    return llave_privada.public_key()
+
+#============= Llaves a byte a pem y viceversa ===================
+
+def convertir_llave_privada_bytes(llave_privada):
+    """
+    Convierte de bytes a PEM
+    """
+    resultado = llave_privada.private_bytes(
+        encoding=serialization.Encoding.PEM,
+        format=serialization.PrivateFormat.TraditionalOpenSSL,
+        encryption_algorithm=serialization.NoEncryption()
+    )
+    return resultado
+
+
+def convertir_bytes_llave_privada(contenido_texto):
+    """
+    Convierte de PEM a bytes
+    """
+    resultado = serialization.load_pem_private_key(
+        contenido_texto,
+        backend=default_backend(),
+        password=None)
+    return resultado
+
+def convertir_llave_publica_bytes(llave_publica):
+    resultado = llave_publica.public_bytes(
+        encoding=serialization.Encoding.PEM,
+        format=serialization.PublicFormat.SubjectPublicKeyInfo
+    )
+    return resultado
+
+
+def convertir_bytes_llave_publica(contenido_texto):
+    resultado = serialization.load_pem_public_key(
+        contenido_texto,
+        backend=default_backend())
+    return resultado
