+++ Proyect_DSW/templates/verificarFirmas.html
@@ -51,7 +51,7 @@
             </div>
             {% endif %}
             {% if confirmacion %}
-            <div id="confirmacion" class="alert alert-danger mt-3">
+            <div id="confirmacion" class="alert alert-success mt-3">
                 <strong>confirmacion:</strong>
                 <ul id="lista-errores" class="mb-0">
                     {% for veri in confirmacion %}
